Remove calculateForgingInfo
